GH-53: Remove version specific initiations and merge tests
