--- tests/antd4.test.tsx
@@ -1,7 +1,0 @@
-import Form from "antd4/lib/form";
-import Button from "antd4/lib/button";
-import FormItem from "antd4/lib/form/FormItem";
-
-import commonTests from "./common";
-
-commonTests(Form, FormItem, Button);
