--- tests/antd5.test.tsx
@@ -1,7 +1,0 @@
-import Form from "antd/lib/form";
-import Button from "antd/lib/button";
-import FormItem from "antd/lib/form/FormItem";
-
-import commonTests from "./common";
-
-commonTests(Form, FormItem, Button);
